Add Sentry logging of fetching errors
